# docs/guides/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Project Manager UI API
  description: |
    REST API for the Project Manager UI — a real-time project management dashboard
    for stakeholders and AI agents. Built with Next.js 15, PostgreSQL 16, Prisma, and SSE.

    **Base URL**: `http://localhost:3000/api` (local) or your deployment host.

    **Authentication** (dual):
    - **Agent access**: `X-API-Key` header (see [ApiKeyAuth](#/components/securitySchemes/ApiKeyAuth))
    - **Browser user access**: `auth_token` cookie set by login/register (see [CookieAuth](#/components/securitySchemes/CookieAuth))
    
    Write operations (POST, PUT, DELETE) require authentication. Read operations (GET) 
    are publicly accessible for backward compatibility (planned enforcement in v2).

    The SSE endpoint (`/api/events`) requires authentication as of v1.1.0.
  version: "1.1.0"
  contact:
    name: Project Manager UI Team
    url: https://github.com/gansru/ui_pm
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for AI agent authentication. Valid for all authenticated operations.
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT cookie for browser user authentication. Set automatically by login/register endpoints. httpOnly, sameSite=lax, 7-day expiry.

    # ─── Request Payloads ───
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - confirmPassword
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Display name (required)
          example: "Roman Govorov"
        email:
          type: string
          format: email
          maxLength: 255
          description: Email address (must not already exist)
          example: "roman@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Password (min 8 characters, cannot be a common password)
          example: "MyStr0ng!P@ss"
        confirmPassword:
          type: string
...
  schemas:
...
    ErrorBody:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_ERROR
            - NOT_FOUND
            - CONFLICT
            - UNAUTHORIZED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                description: Dot-path of the invalid field
              message:
                type: string
                description: Validation error message
          description: Detailed validation errors (present only for VALIDATION_ERROR)
...
    PaginatedResponseProject:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
          description: Total number of records returned
...
    Project:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique project identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          maxLength: 255
          description: Project name
          example: "AI Education Platform"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Optional project description
          example: "Building an AI-powered education tool"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
        _count:
          type: object
          description: Related entity counts
          properties:
            tasks:
              type: integer
        subprojects:
          type: array
          items:
            $ref: "#/components/schemas/Subproject"
          description: Associated subprojects (returned by GET /{id} only)
    Subproject:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Task:
      type: object
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - assignee
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        projectId:
          type: string
          format: uuid
          description: Associated project ID
        subprojectId:
          type: string
          format: uuid
          nullable: true
          description: Associated subproject ID (optional)
        title:
          type: string
          maxLength: 200
          description: Task title
          example: "Implement login page"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Detailed task description
        status:
          type: string
          enum:
            - in_work
            - review
            - done
          description: Current workflow status
          example: "in_work"
        priority:
          type: string
          enum:
            - high
            - medium
            - low
          description: Task priority level
          example: "medium"
        assignee:
          type: string
          maxLength: 100
          description: Name of assigned user
          example: "Roman Govorov"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
...
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Project name (required)
          example: "New Initiative"
        description:
          type: string
          maxLength: 2000
          description: Optional project description
          example: "Building something great"
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated project name
          example: "Renamed Initiative"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Updated or cleared description
          example: "Updated description"
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated title
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Updated or cleared description
        status:
          type: string
          enum:
            - in_work
            - review
            - done
          description: Updated status
        priority:
          type: string
          enum:
            - high
            - medium
            - low
          description: Updated priority
        assignee:
          type: string
          minLength: 1
          maxLength: 100
          description: Updated assignee name
        projectId:
          type: string
          format: uuid
          description: Updated project assignment (valid UUID)
        subprojectId:
          type: string
          format: uuid
          nullable: true
          description: Updated subproject assignment
...
paths:
...
  /projects:
    get:
      tags:
        - projects
      summary: List all projects
      description: |
        Returns a paginated list of all projects, ordered alphabetically by name.
        Each project includes a `_count.tasks` field showing the number of associated tasks.
      operationId: listProjects
      responses:
        "200":
          description: A list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponseProject"
              example:
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "AI Education Platform"
                    description: "Building an AI-powered education tool"
                    createdAt: "2026-08-01T10:00:00.000Z"
                    updatedAt: "2026-08-10T14:30:00.000Z"
                    _count: { tasks: 12 }
                total: 1
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags:
        - projects
      summary: Create a new project
      description: |
        Creates a new project with the given name and optional description.
        Requires API key authentication. Emits a `project_created` SSE event.
      operationId: createProject
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              name: "New Initiative"
              description: "A brand new project"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Unique constraint violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /projects/{id}:
    get:
      tags:
        - projects
      summary: Get project by ID
      description: |
        Returns a single project by its UUID. Includes subprojects and task count.
        No authentication required.
      operationId: getProjectById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the project
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project details including subprojects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    put:
      tags:
        - projects
      summary: Update a project
      description: |
        Updates existing project fields. Only provided fields are modified.
        Partial updates are supported (omit fields you don't want to change).
        Requires API key authentication. Emits a `project_updated` SSE event.
      operationId: updateProject
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the project to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
            example:
              name: "Updated Project Name"
              description: "New description text"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

    delete:
      tags:
        - projects
      summary: Delete a project
      description: |
        Permanently deletes a project and all its associated tasks.
        Requires API key authentication. Emits a `project_deleted` SSE event.
      operationId: deleteProject
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the project to delete
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project deleted successfully (returns `{ id }`)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Foreign key constraint (tasks reference this project)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ============================================================
  # TASKS — /api/tasks
  # ============================================================
...
  /tasks/{id}:
    get:
      tags:
        - tasks
      summary: Get task by ID
      description: |
        Returns a single task by its UUID. Includes full details.
        No authentication required.
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    put:
      tags:
        - tasks
      summary: Update a task
      description: |
        Updates existing task fields. Only provided fields are modified.
        Partial updates are supported. Status transitions follow the Kanban flow:
        `in_work` → `review` → `done`.
        Requires API key authentication. Emits a `task_updated` SSE event.
      operationId: updateTask
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
            example:
              status: review
              priority: high
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error or empty update payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

    delete:
      tags:
        - tasks
      summary: Delete a task
      description: |
        Permanently deletes a task. Requires API key authentication.
        Emits a `task_deleted` SSE event.
      operationId: deleteTask
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to delete
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task deleted successfully (returns `{ id }`)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ============================================================
  # SUBTASKS — /api/projects/{id}/tasks
  # ============================================================
...